requirements.txt
flask==3.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai>=1.12.0
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
requests==2.31.0
httpx>=0.27.0
